refactor: ability service interface 생성
