+++ src/main/java/corecord/dev/domain/ability/domain/repository/AbilityRepository.java
@@ -19,10 +19,10 @@ public interface AbilityRepository extends JpaRepository<Ability, Long> {
                 "a.keyword, COUNT(a), " + // 각 키워드의 개수 집계
                 "(COUNT(a) * 1.0 / (SELECT COUNT(a2) FROM Ability a2 WHERE a2.user = :user)) * 100.0) " + // 각 키워드의 비율 집계
                 "FROM Ability a " +
-                "WHERE a.user = :user " +
+                "WHERE a.user.userId = :userId " +
                 "GROUP BY a.keyword " +
                 "ORDER BY 3 desc ") // 비율 높은 순 정렬
-        List<AbilityResponse.KeywordStateDto> findKeywordStateDtoList(@Param(value = "user") User user);
+        List<AbilityResponse.KeywordStateDto> findKeywordStateDtoList(@Param(value = "userId") Long userId);
 
         @Modifying
         @Query("DELETE " +
@@ -39,8 +39,8 @@ public interface AbilityRepository extends JpaRepository<Ability, Long> {
         @Query("SELECT distinct a.keyword AS keyword " + // unique한 keyword list 반환
                 "FROM Ability a " +
                 "JOIN a.analysis ana " +
-                "WHERE a.user = :user " +
+                "WHERE a.user.userId = :userId " +
                 "GROUP BY a.keyword " +
                 "ORDER BY COUNT(a.keyword) DESC, MAX(ana.createdAt) DESC ") // 개수가 많은 순, 최근 생성 순 정렬
-        List<Keyword> getKeywordList(@Param(value = "user") User user);
+        List<Keyword> getKeywordList(@Param(value = "userId") Long userId);
 }
